apps/api/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
aiohttp
influxdb-client
pymongo
SQLAlchemy[asyncio]
google-generativeai
anthropic
pytest
pydantic-settings
aiosqlite
python-dotenv
structlog
rich
alembic
asyncpg
psycopg2-binary
cryptography
celery
redis
websockets
greenlet
